feat: add constant in param PDO asynchrone request

+++ Api/delete-review.php
@@ -3,13 +3,20 @@ require_once $_SERVER['DOCUMENT_ROOT'] . '/config.php';
 
 $data = json_decode(file_get_contents('php://input'), true);
 
+if (!isset($data['review_id'])) {
+    http_response_code(400);
+    echo json_encode(['error' => 'Missing required fields']);
+    exit;
+}
+
 $reviewId = (int)$data['review_id'];
 
 try {
-    $pdo = new PDO("mysql:dbname=" . _DB_NAME_ . ";host=db;charset=utf8mb4", _DB_USER_, _DB_PASSWORD_);
+    $pdo = new PDO("mysql:dbname=" . _DB_NAME_ . ";host=" . _DB_HOST_ . ";charset=utf8mb4", _DB_USER_, _DB_PASSWORD_);
     $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 
     $stmt = $pdo->prepare('DELETE FROM review WHERE id = :id');
+
     $stmt->bindValue(":id", $reviewId, PDO::PARAM_INT);
     $result = $stmt->execute();
